Move Arguments to their own Model\ProbeArgument namespace.
