+++ src/Model/ProbeArgument/HttpArguments.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App;
+namespace App\Model\ProbeArgument;
 
 class HttpArguments extends ProbeArguments
 {
